move the provider into context folder
